refactor(layout): move footer from sections to layout
Both are global chrome, not page sections.
